refactor(backend): break up monolithic main.go into modular internal/handlers
